Update patch set 2
Patch Set 2:

I would like you to apply 3 patches to solve the problem with the camera on the shinano platform. I know I copied these patches from gerrit (thank's koron393), but I split the build for rhine-common with the unmodified patch, and for shinano-common did with the applied patches. I'm sorry, I don't know how to add authorship to my patches, I ask you to create changes on gerrit and add corresponding authors and me as the one who adapted this patch. Check the patches in advance.
1. https://github.com/Raschad/android_device_sony_msm8974-common/commit/783ebc6aa443650714b83fa6556e9d6bf21168c1
2. https://github.com/Raschad/device_sony_rhine-common/commit/57b8c4d68475fa4f6ba657ab02156b4213efa9f7
3. https://github.com/Raschad/device_sony_shinano-common/commit/69f00869f13ff63d8658bfd58980410bdc83756a

Patch-set: 2
CC: Gerrit User 335 <335@4c0b52a2-b090-4331-a190-b288d812d081>